:recycle: 웹소켓에서 kafka는 비동기로 처리하도록 수정
