+++ src/main/kotlin/com/stark/shoot/application/port/in/chat/SendMessageUseCase.kt
@@ -1,7 +1,8 @@
 package com.stark.shoot.application.port.`in`.chat
 
 import com.stark.shoot.adapter.`in`.web.dto.ChatMessageRequest
+import java.util.concurrent.CompletableFuture
 
 interface SendMessageUseCase {
-    fun handleMessage(message: ChatMessageRequest)
+    fun handleMessage(message: ChatMessageRequest): CompletableFuture<Void>
 }
