+++ src/main/kotlin/com/stark/shoot/application/service/chat/SendMessageService.kt
@@ -10,13 +10,14 @@ import com.stark.shoot.domain.chat.event.EventType
 import com.stark.shoot.domain.chat.message.ChatMessage
 import com.stark.shoot.domain.chat.message.MessageContent
 import org.springframework.stereotype.Service
+import java.util.concurrent.CompletableFuture
 
 @Service
 class SendMessageService(
     private val kafkaMessagePublishPort: KafkaMessagePublishPort
 ) : SendMessageUseCase {
 
-    override fun handleMessage(message: ChatMessageRequest) {
+    override fun handleMessage(message: ChatMessageRequest): CompletableFuture<Void> {
         // ChatMessage 생성
         val chatMessage = ChatMessage(
             roomId = message.roomId,
@@ -34,8 +35,8 @@ class SendMessageService(
             data = chatMessage
         )
 
-        // kafka로 이벤트 발행
-        kafkaMessagePublishPort.publishChatEvent(
+        // kafka로 이벤트 발행하고 CompletableFuture 반환
+        return kafkaMessagePublishPort.publishChatEvent(
             topic = "chat-messages",
             key = message.roomId,
             event = chatEvent
